+++ lib/levanter/src/levanter/grug/grug_moe.py
@@ -1,7 +1,7 @@
 # Copyright The Levanter Authors
 # SPDX-License-Identifier: Apache-2.0
 
-"""Canonical compact Grug MoE kernels.
+"""Public Grug MoE interface and implementation dispatcher.
 
 Implementation overview:
 - Routing keeps the argsort-grouped dispatch path that emerged as the stable
@@ -10,144 +10,52 @@ Implementation overview:
 - Expert parallelism keeps the ring-style strategy from
   https://github.com/marin-community/marin/issues/2710: token-sharded
   `all_gather` for dispatch, then `psum_scatter` for collection.
-- This module intentionally provides functional kernels only; model/module
-  wiring lives in the Grug model files.
+- Backend bodies live in focused `moe_*` modules; this module keeps the stable
+  public API used by Grug model code and benchmarks.
 """
 
-import math
-
 from collections.abc import Callable
-from dataclasses import dataclass
 from functools import partial
-from typing import Literal, TypeAlias, cast, get_args
 
 import equinox as eqx
 import jax
 import jax.numpy as jnp
-from haliax.jax_utils import named_call, tree_checkpoint_name
+from haliax.jax_utils import named_call
 from jax import shard_map
-from jax.sharding import Mesh, NamedSharding, PartitionSpec as P, get_abstract_mesh, get_mesh, reshard
-from jaxtyping import Array, Bool, Float, Int
+from jax.sharding import PartitionSpec as P
+from jaxtyping import Array, Float, Int
 
-from haliax.nn.ragged_dot import ragged_dot
+from levanter.grug.moe_common import (
+    _CUSTOM_VJP_DOWN_MOE_IMPLEMENTATIONS,
+    _DEFAULT_EP_CAPACITY_FACTOR,
+    _EP_MOE_IMPLEMENTATIONS,
+    _batch_spec_from_x,
+    _current_mesh,
+    _init_weight,
+    _mesh_axis_size,
+    _mesh_has_axis,
+    _reshard_for_init,
+    _reshard_for_shard_map,
+    _value_spec_or_default,
+    interleave_moe_w13,
+    moe_implementation_uses_interleaved_w13,
+    MoEExpertMlpPspecs,
+    MoeActivation,
+    MoeImplementation,
+    PspecAxis,
+    resolve_moe_implementation,
+    split_moe_w13_output,
+)
+from levanter.grug.moe_ep_common import (
+    _clip_receiver_group_sizes as _clip_receiver_group_sizes,
+    _compact_by_keep_mask as _compact_by_keep_mask,
+    _expand_from_keep_mask as _expand_from_keep_mask,
+    _shard_a2a_params as _shard_a2a_params,
+)
+from levanter.grug.moe_ep_ragged_all_to_all import _moe_mlp_ep_ragged_a2a_local
+from levanter.grug.moe_ep_ring import _moe_mlp_ep_ring_local
+from levanter.grug.moe_local import _moe_mlp_local
 from levanter.utils.activation import ActivationFunctionEnum
-
-_DEFAULT_EP_CAPACITY_FACTOR = 1.25
-# #2710 used 1.25 as the practical EP ring default to avoid over/under-packing.
-
-PspecAxis: TypeAlias = str | tuple[str, ...] | None
-MoeActivation: TypeAlias = ActivationFunctionEnum | Callable[[jax.Array], jax.Array]
-MoeImplementation: TypeAlias = Literal[
-    "ring",
-    "ragged_all_to_all",
-    "scatter",
-    "sonic_xla",
-    "sonic_xla_interleaved_w13",
-    "sonic_xla_interleaved_w13_custom_vjp_down",
-]
-_VALID_MOE_IMPLEMENTATIONS = get_args(MoeImplementation)
-_EP_MOE_IMPLEMENTATIONS = ("ring", "ragged_all_to_all")
-_LOCAL_MOE_IMPLEMENTATIONS = (
-    "scatter",
-    "sonic_xla",
-    "sonic_xla_interleaved_w13",
-    "sonic_xla_interleaved_w13_custom_vjp_down",
-)
-_INTERLEAVED_W13_MOE_IMPLEMENTATIONS = (
-    "sonic_xla_interleaved_w13",
-    "sonic_xla_interleaved_w13_custom_vjp_down",
-)
-_CUSTOM_VJP_DOWN_MOE_IMPLEMENTATIONS = ("sonic_xla_interleaved_w13_custom_vjp_down",)
-
-
-@dataclass(frozen=True)
-class MoEExpertMlpPspecs:
-    """Logical sharding axes for local MoE expert MLP weights."""
-
-    expert: PspecAxis = "expert"
-    hidden: PspecAxis = "data"
-    intermediate: PspecAxis = "model"
-
-    @property
-    def w_gate_up(self) -> P:
-        return P(self.expert, self.hidden, self.intermediate)
-
-    @property
-    def w_down(self) -> P:
-        return P(self.expert, self.intermediate, self.hidden)
-
-
-def _current_mesh() -> Mesh | jax.sharding.AbstractMesh:
-    try:
-        mesh = get_mesh()
-    except ValueError:
-        mesh = None
-    if mesh is not None and not mesh.empty:
-        return mesh
-    return get_abstract_mesh()
-
-
-def _mesh_has_axis(mesh: Mesh | jax.sharding.AbstractMesh | None, axis_name: str) -> bool:
-    if mesh is None or mesh.empty:
-        return False
-    return axis_name in mesh.shape
-
-
-def _mesh_axis_size(mesh: Mesh | jax.sharding.AbstractMesh | None, axis_name: str) -> int:
-    if mesh is None or mesh.empty:
-        return 1
-    return int(mesh.shape.get(axis_name, 1))
-
-
-def _batch_spec(mesh: Mesh | jax.sharding.AbstractMesh | None) -> P:
-    if _mesh_has_axis(mesh, "expert"):
-        return P(("data", "expert"))
-    return P(("data",))
-
-
-def _reshard_for_init(x: jax.Array, spec: P) -> jax.Array:
-    mesh = _current_mesh()
-    if mesh is None or mesh.empty:
-        return x
-    return reshard(x, NamedSharding(mesh, spec))
-
-
-def resolve_moe_implementation(implementation: MoeImplementation | str | None) -> MoeImplementation:
-    if implementation is None:
-        return "ring"
-    if implementation not in _VALID_MOE_IMPLEMENTATIONS:
-        valid = ", ".join(repr(choice) for choice in _VALID_MOE_IMPLEMENTATIONS)
-        raise ValueError(f"implementation must be one of {valid} or None, got {implementation!r}")
-    return cast(MoeImplementation, implementation)
-
-
-def moe_implementation_uses_interleaved_w13(
-    implementation: MoeImplementation | str | None,
-) -> bool:
-    return resolve_moe_implementation(implementation) in _INTERLEAVED_W13_MOE_IMPLEMENTATIONS
-
-
-def interleave_moe_w13(w_gate: jax.Array, w_up: jax.Array) -> jax.Array:
-    """Pack concat-style gate/up expert weights into Sonic's interleaved GLU layout."""
-    if w_gate.shape != w_up.shape:
-        raise ValueError(f"w_gate and w_up must have the same shape, got {w_gate.shape} vs {w_up.shape}")
-    return jnp.stack((w_gate, w_up), axis=-1).reshape(*w_gate.shape[:-1], 2 * w_gate.shape[-1])
-
-
-def split_moe_w13_output(
-    w13_out: jax.Array, *, intermediate_dim: int, interleaved: bool
-) -> tuple[jax.Array, jax.Array]:
-    expected = 2 * intermediate_dim
-    if w13_out.shape[-1] != expected:
-        raise ValueError(f"w13 output last dimension must be {expected}, got shape={w13_out.shape}")
-    if interleaved:
-        return w13_out[..., 0::2], w13_out[..., 1::2]
-    gate, up = jnp.split(w13_out, [intermediate_dim], axis=-1)
-    return gate, up
-
-
-def _init_weight(key: jax.Array, shape: tuple[int, ...], std: float) -> Float[Array, "..."]:
-    return std * jax.random.truncated_normal(key, -3, 3, shape)
 
 
 class MoEExpertMlp(eqx.Module):
@@ -216,669 +124,6 @@ class MoEExpertMlp(eqx.Module):
         )
 
 
-def _custom_vjp_interleaved_down_gather_sum(
-    w13_out_interleaved: jax.Array,
-    combine_weights: jax.Array,
-    moe_w2: jax.Array,
-    token_ids_sort: jax.Array,
-    sorted_assignment_ids: jax.Array,
-    dispatch_positions: jax.Array,
-    group_sizes: jax.Array,
-) -> jax.Array:
-    from levanter.grug.custom_vjp_moe import custom_vjp_interleaved_down_gather_sum
-
-    return custom_vjp_interleaved_down_gather_sum(
-        w13_out_interleaved,
-        combine_weights,
-        moe_w2,
-        token_ids_sort,
-        sorted_assignment_ids,
-        dispatch_positions,
-        group_sizes,
-    )
-
-
-@named_call
-def _prepare_moe_dispatch(
-    x: Float[Array, "T D"],
-    selected_experts: Int[Array, "T K"],
-    combine_weights: Float[Array, "T K"],
-    *,
-    num_experts: int,
-) -> tuple[
-    Float[Array, "TK D"],
-    Float[Array, "TK"],
-    Int[Array, "TK"],
-    Int[Array, "E"],
-]:
-    """Flatten + argsort by expert into grouped layout for GMM."""
-    # #2704: keep argsort-grouped dispatch as the canonical compact routing
-    # strategy, matching the behavior carried forward from 89318a910.
-    tokens, topk = selected_experts.shape
-    expert_ids = selected_experts.reshape(tokens * topk)
-    dispatch_weights = combine_weights.reshape(tokens * topk)
-
-    sort_idx = jnp.argsort(expert_ids, axis=0)
-    token_ids = jnp.arange(tokens * topk, dtype=jnp.int32) // topk
-    token_ids_sort = token_ids[sort_idx]
-    x_sort = x[token_ids_sort]
-    w_sort = dispatch_weights[sort_idx].astype(x.dtype)
-    group_sizes = jnp.bincount(expert_ids, length=num_experts).astype(jnp.int32)
-    return x_sort, w_sort, token_ids_sort, group_sizes
-
-
-@named_call
-def _prepare_moe_dispatch_indices_with_assignment_ids(
-    selected_experts: Int[Array, "T K"],
-    *,
-    num_experts: int,
-) -> tuple[
-    Int[Array, "TK"],
-    Int[Array, "T K"],
-    Int[Array, "E"],
-    Int[Array, "TK"],
-]:
-    """Prepare expert-sorted token ids plus reverse positions without gathering x."""
-    tokens, topk = selected_experts.shape
-    assignments = tokens * topk
-    expert_ids = selected_experts.reshape(assignments)
-
-    sort_idx = jnp.argsort(expert_ids, axis=0)
-    assignment_ids = jnp.arange(assignments, dtype=jnp.int32)
-    sorted_assignment_ids = assignment_ids[sort_idx]
-    token_ids_sort = sorted_assignment_ids // topk
-
-    sorted_positions = jnp.arange(assignments, dtype=jnp.int32)
-    dispatch_positions = jnp.zeros((assignments,), dtype=jnp.int32).at[sort_idx].set(sorted_positions)
-    dispatch_positions = dispatch_positions.reshape(tokens, topk)
-
-    group_sizes = jnp.bincount(expert_ids, length=num_experts).astype(jnp.int32)
-    return token_ids_sort, dispatch_positions, group_sizes, sorted_assignment_ids
-
-
-def _gather_sum_reference(
-    dispatch_output: Float[Array, "TK D"],
-    dispatch_positions: Int[Array, "T K"],
-    combine_weights: Float[Array, "T K"],
-) -> Float[Array, "T D"]:
-    acc = jnp.zeros((dispatch_positions.shape[0], dispatch_output.shape[1]), dtype=dispatch_output.dtype)
-    weights = combine_weights.astype(dispatch_output.dtype)
-    for topk_index in range(dispatch_positions.shape[1]):
-        gathered = jnp.take(dispatch_output, dispatch_positions[:, topk_index], axis=0)
-        acc = (acc + (gathered * weights[:, topk_index, None]).astype(dispatch_output.dtype)).astype(
-            dispatch_output.dtype
-        )
-    return acc
-
-
-def _zero_dropped_assignments() -> Int[Array, ""]:
-    return jnp.array(0, dtype=jnp.int32)
-
-
-def _moe_mlp_local_scatter(
-    x: Float[Array, "T D"],
-    selected_experts: Int[Array, "T K"],
-    combine_weights: Float[Array, "T K"],
-    moe_w13: Float[Array, "E D I2"],
-    moe_w2: Float[Array, "E I D"],
-    *,
-    activation_fn: Callable[[jax.Array], jax.Array],
-    num_experts: int,
-) -> tuple[Float[Array, "T D"], Int[Array, ""]]:
-    """Local fallback MoE path: sorted grouped GMM then scatter-add combine."""
-    x_dispatch, w_dispatch, token_dispatch, group_sizes = _prepare_moe_dispatch(
-        x,
-        selected_experts,
-        combine_weights,
-        num_experts=num_experts,
-    )
-    x_dispatch = tree_checkpoint_name(x_dispatch, "grug_moe_dispatch_input")
-
-    with jax.named_scope("moe_up_down"):
-        w13_out = tree_checkpoint_name(ragged_dot(x_dispatch, moe_w13, group_sizes), "grug_moe_expert_hidden")
-        moe_dim = moe_w2.shape[1]
-        gate, up = split_moe_w13_output(w13_out, intermediate_dim=moe_dim, interleaved=False)
-        out_dispatch = tree_checkpoint_name(
-            ragged_dot(activation_fn(gate) * up, moe_w2, group_sizes),
-            "grug_moe_dispatch_output",
-        )
-
-    with jax.named_scope("scatter"):
-        out = jnp.zeros_like(x).at[token_dispatch].add(out_dispatch * w_dispatch[:, None], mode="drop")
-    return out, _zero_dropped_assignments()
-
-
-def _moe_mlp_local_sonic_xla(
-    x: Float[Array, "T D"],
-    selected_experts: Int[Array, "T K"],
-    combine_weights: Float[Array, "T K"],
-    moe_w13: Float[Array, "E D I2"],
-    moe_w2: Float[Array, "E I D"],
-    *,
-    activation_fn: Callable[[jax.Array], jax.Array],
-    num_experts: int,
-) -> tuple[Float[Array, "T D"], Int[Array, ""]]:
-    """Local Sonic-style path with XLA gather-sum combine and concat W13 layout."""
-    token_ids_sort, dispatch_positions, group_sizes, _sorted_assignment_ids = (
-        _prepare_moe_dispatch_indices_with_assignment_ids(
-            selected_experts,
-            num_experts=num_experts,
-        )
-    )
-    x_dispatch = tree_checkpoint_name(x[token_ids_sort], "grug_moe_dispatch_input")
-
-    with jax.named_scope("moe_up_down"):
-        w13_out = tree_checkpoint_name(ragged_dot(x_dispatch, moe_w13, group_sizes), "grug_moe_expert_hidden")
-        moe_dim = moe_w2.shape[1]
-        gate, up = split_moe_w13_output(w13_out, intermediate_dim=moe_dim, interleaved=False)
-        out_dispatch = tree_checkpoint_name(
-            ragged_dot(activation_fn(gate) * up, moe_w2, group_sizes),
-            "grug_moe_dispatch_output",
-        )
-
-    with jax.named_scope("gather_sum"):
-        out = _gather_sum_reference(out_dispatch, dispatch_positions, combine_weights)
-    return out, _zero_dropped_assignments()
-
-
-def _moe_mlp_local_sonic_xla_interleaved_w13(
-    x: Float[Array, "T D"],
-    selected_experts: Int[Array, "T K"],
-    combine_weights: Float[Array, "T K"],
-    moe_w13: Float[Array, "E D I2"],
-    moe_w2: Float[Array, "E I D"],
-    *,
-    activation_fn: Callable[[jax.Array], jax.Array],
-    num_experts: int,
-) -> tuple[Float[Array, "T D"], Int[Array, ""]]:
-    """Local Sonic-style path with XLA gather-sum combine and interleaved W13 layout."""
-    token_ids_sort, dispatch_positions, group_sizes, _sorted_assignment_ids = (
-        _prepare_moe_dispatch_indices_with_assignment_ids(
-            selected_experts,
-            num_experts=num_experts,
-        )
-    )
-    x_dispatch = tree_checkpoint_name(x[token_ids_sort], "grug_moe_dispatch_input")
-
-    with jax.named_scope("moe_up_down"):
-        w13_out = tree_checkpoint_name(ragged_dot(x_dispatch, moe_w13, group_sizes), "grug_moe_expert_hidden")
-        moe_dim = moe_w2.shape[1]
-        gate, up = split_moe_w13_output(w13_out, intermediate_dim=moe_dim, interleaved=True)
-        out_dispatch = tree_checkpoint_name(
-            ragged_dot(activation_fn(gate) * up, moe_w2, group_sizes),
-            "grug_moe_dispatch_output",
-        )
-
-    with jax.named_scope("gather_sum"):
-        out = _gather_sum_reference(out_dispatch, dispatch_positions, combine_weights)
-    return out, _zero_dropped_assignments()
-
-
-def _moe_mlp_local_sonic_xla_interleaved_w13_custom_vjp_down(
-    x: Float[Array, "T D"],
-    selected_experts: Int[Array, "T K"],
-    combine_weights: Float[Array, "T K"],
-    moe_w13: Float[Array, "E D I2"],
-    moe_w2: Float[Array, "E I D"],
-    *,
-    activation_fn: Callable[[jax.Array], jax.Array],
-    num_experts: int,
-) -> tuple[Float[Array, "T D"], Int[Array, ""]]:
-    """Local Sonic-style path with XLA W13 and a custom-VJP down/gather boundary."""
-    del activation_fn
-    token_ids_sort, dispatch_positions, group_sizes, sorted_assignment_ids = (
-        _prepare_moe_dispatch_indices_with_assignment_ids(
-            selected_experts,
-            num_experts=num_experts,
-        )
-    )
-    x_dispatch = tree_checkpoint_name(x[token_ids_sort], "grug_moe_dispatch_input")
-
-    with jax.named_scope("moe_up_down"):
-        w13_out = tree_checkpoint_name(ragged_dot(x_dispatch, moe_w13, group_sizes), "grug_moe_expert_hidden")
-        out = tree_checkpoint_name(
-            _custom_vjp_interleaved_down_gather_sum(
-                w13_out,
-                combine_weights,
-                moe_w2,
-                token_ids_sort,
-                sorted_assignment_ids,
-                dispatch_positions,
-                group_sizes,
-            ),
-            "grug_moe_output",
-        )
-    return out, _zero_dropped_assignments()
-
-
-_MOE_LOCAL_FNS = {
-    "scatter": _moe_mlp_local_scatter,
-    "sonic_xla": _moe_mlp_local_sonic_xla,
-    "sonic_xla_interleaved_w13": _moe_mlp_local_sonic_xla_interleaved_w13,
-    "sonic_xla_interleaved_w13_custom_vjp_down": _moe_mlp_local_sonic_xla_interleaved_w13_custom_vjp_down,
-}
-
-
-def _moe_mlp_local(
-    x: Float[Array, "T D"],
-    selected_experts: Int[Array, "T K"],
-    combine_weights: Float[Array, "T K"],
-    moe_w13: Float[Array, "E D I2"],
-    moe_w2: Float[Array, "E I D"],
-    *,
-    activation_fn: Callable[[jax.Array], jax.Array],
-    num_experts: int,
-    implementation: MoeImplementation,
-) -> tuple[Float[Array, "T D"], Int[Array, ""]]:
-    local_key = implementation if implementation in _LOCAL_MOE_IMPLEMENTATIONS else "scatter"
-    return _MOE_LOCAL_FNS[local_key](
-        x,
-        selected_experts,
-        combine_weights,
-        moe_w13,
-        moe_w2,
-        activation_fn=activation_fn,
-        num_experts=num_experts,
-    )
-
-
-def _batch_spec_from_x(x: jax.Array, mesh: Mesh | jax.sharding.AbstractMesh | None) -> P:
-    sharding = getattr(x, "sharding", None)
-    spec = getattr(sharding, "spec", None)
-    if spec is not None and len(spec) > 0 and spec[0] is not None:
-        return P(spec[0])
-    return _batch_spec(mesh)
-
-
-def _is_replicated_spec(spec: P) -> bool:
-    return all(axis is None for axis in spec)
-
-
-def _value_spec_or_default(x: jax.Array, default: P, *, replace_replicated: bool = False) -> P:
-    sharding = getattr(x, "sharding", None)
-    spec = getattr(sharding, "spec", None)
-    if spec is not None and not (replace_replicated and _is_replicated_spec(spec)):
-        return spec
-    return default
-
-
-def _reshard_for_shard_map(x: jax.Array, mesh: Mesh | jax.sharding.AbstractMesh | None, spec: P) -> jax.Array:
-    if mesh is not None and not mesh.empty:
-        return reshard(x, NamedSharding(mesh, spec))
-    return x
-
-
-def _sort_activations(inputs: jax.Array, sort_indices: jax.Array) -> jax.Array:
-    if inputs.shape[0] != sort_indices.shape[0]:
-        raise ValueError(f"Expected matching leading dims, got {inputs.shape[0]} and {sort_indices.shape[0]}")
-    return _sort_activations_custom(inputs, sort_indices)
-
-
-@jax.custom_vjp
-def _sort_activations_custom(inputs: jax.Array, sort_indices: jax.Array) -> jax.Array:
-    return inputs[sort_indices, ...]
-
-
-def _sort_activations_custom_fwd(inputs: jax.Array, sort_indices: jax.Array) -> tuple[jax.Array, jax.Array]:
-    return _sort_activations_custom(inputs, sort_indices), sort_indices
-
-
-def _sort_activations_custom_bwd(residuals: jax.Array, grads: jax.Array) -> tuple[jax.Array, None]:
-    sort_indices = residuals
-    return _sort_activations_custom(grads, jnp.argsort(sort_indices)), None
-
-
-_sort_activations_custom.defvjp(_sort_activations_custom_fwd, _sort_activations_custom_bwd)
-
-
-def _prefix_cap_counts(counts: Int[Array, "E"], *, capacity: int) -> Int[Array, "E"]:
-    accepted = []
-    remaining = jnp.array(capacity, dtype=jnp.int32)
-    for expert in range(int(counts.shape[0])):
-        take = jnp.minimum(counts[expert], remaining)
-        accepted.append(take)
-        remaining = jnp.maximum(remaining - take, 0)
-    return jnp.stack(accepted, axis=0)
-
-
-def _permute_by_global_expert(
-    x_local: jax.Array,
-    selected_experts_local: jax.Array,
-    *,
-    num_experts: int,
-) -> tuple[jax.Array, jax.Array, jax.Array]:
-    topk = selected_experts_local.shape[1]
-    flat_selected = selected_experts_local.reshape(-1)
-    sorted_indices = jnp.argsort(flat_selected)
-    repeated_x = jnp.repeat(x_local, topk, axis=0)
-    sorted_x = _sort_activations(repeated_x, sorted_indices)
-    group_sizes = jnp.bincount(flat_selected, length=num_experts).astype(jnp.int32)
-    return sorted_x, sorted_indices, group_sizes
-
-
-def _unpermute_from_global_expert(
-    intermediate: jax.Array,
-    sorted_indices: jax.Array,
-    combine_weights_local: jax.Array,
-    *,
-    tokens_per_shard: int,
-    topk: int,
-) -> jax.Array:
-    unsorted = _sort_activations(intermediate, jnp.argsort(sorted_indices))
-    reshaped = unsorted.reshape(tokens_per_shard, topk, -1)
-    return jnp.einsum(
-        "tkd,tk->td", reshaped, combine_weights_local.astype(reshaped.dtype), preferred_element_type=jnp.float32
-    )
-
-
-def _shard_a2a_params(
-    shard_counts: jax.Array,
-    shard_id: jax.Array,
-) -> tuple[jax.Array, jax.Array, jax.Array, jax.Array]:
-    row = shard_counts[shard_id]
-    input_offsets = jnp.cumsum(jnp.concatenate((jnp.array([0], dtype=row.dtype), row[:-1])))
-    send_sizes = row
-
-    recv_sizes = shard_counts[:, shard_id]
-    # `ragged_all_to_all` expects sender-side output offsets: for each
-    # destination shard, where this sender's slice should land in the remote
-    # receiver buffer. JAX computes the local receive offsets by transposing
-    # these offsets with an internal all_to_all.
-    sender_output_offsets = jnp.cumsum(shard_counts, axis=0, dtype=shard_counts.dtype) - shard_counts
-    output_offsets = sender_output_offsets[shard_id]
-    return input_offsets, send_sizes, output_offsets, recv_sizes
-
-
-def _local_permute_from_counts(
-    inputs: jax.Array,
-    global_group_sizes: jax.Array,
-    *,
-    local_expert_size: int,
-    shard_index: jax.Array,
-) -> tuple[jax.Array, jax.Array, jax.Array]:
-    all_shard_local_sizes = jax.lax.dynamic_slice_in_dim(
-        global_group_sizes,
-        start_index=shard_index * local_expert_size,
-        slice_size=local_expert_size,
-        axis=1,
-    )
-    local_group_sizes = jnp.sum(all_shard_local_sizes, axis=0)
-    local_sizes = all_shard_local_sizes.reshape(-1)
-    total_valid = jnp.sum(local_sizes, dtype=jnp.int32)
-    segment_ends = jnp.cumsum(local_sizes, dtype=jnp.int32)
-    positions = jnp.arange(inputs.shape[0], dtype=jnp.int32)
-    segment_index = jnp.searchsorted(segment_ends, positions, side="right")
-    local_expert_ids = jnp.where(positions < total_valid, segment_index % local_expert_size, local_expert_size)
-    sorted_indices = jnp.argsort(local_expert_ids)
-    sorted_inputs = _sort_activations(inputs, sorted_indices)
-    sorted_inputs = jnp.where((positions < total_valid)[:, None], sorted_inputs, 0)
-    group_sizes = local_group_sizes.at[-1].add(inputs.shape[0] - total_valid)
-    return sorted_inputs, sorted_indices, group_sizes
-
-
-def _clip_receiver_group_sizes(
-    global_group_sizes: Int[Array, "S E"],
-    *,
-    local_expert_size: int,
-    receiver_capacity: int,
-) -> Int[Array, "S E"]:
-    """Clip sender->expert group sizes so each receiver shard stays within capacity."""
-    num_senders = int(global_group_sizes.shape[0])
-    num_experts = int(global_group_sizes.shape[1])
-    if num_experts % local_expert_size != 0:
-        raise ValueError(f"num_experts={num_experts} must be divisible by local_expert_size={local_expert_size}")
-    num_receivers = num_experts // local_expert_size
-    if num_receivers != num_senders:
-        raise ValueError(f"sender/receiver shard mismatch: num_senders={num_senders}, num_receivers={num_receivers}")
-
-    clipped_by_receiver: list[jax.Array] = []
-    for receiver_index in range(num_receivers):
-        start = receiver_index * local_expert_size
-        stop = start + local_expert_size
-        receiver_counts = global_group_sizes[:, start:stop]
-        receiver_totals = jnp.sum(receiver_counts, axis=0, dtype=jnp.int32)
-        accepted_totals = _prefix_cap_counts(receiver_totals, capacity=receiver_capacity)
-        remaining = accepted_totals
-        accepted_rows: list[jax.Array] = []
-        for sender_index in range(num_senders):
-            # Greedy first-sender-wins: earlier shards get priority when capacity is scarce.
-            accepted = jnp.minimum(receiver_counts[sender_index], remaining)
-            accepted_rows.append(accepted)
-            remaining = remaining - accepted
-        clipped_by_receiver.append(jnp.stack(accepted_rows, axis=0))
-
-    return jnp.concatenate(clipped_by_receiver, axis=1)
-
-
-def _expert_prefix_keep_mask(
-    group_sizes: Int[Array, "E"],
-    accepted_group_sizes: Int[Array, "E"],
-    *,
-    total_size: int,
-) -> Bool[Array, "T"]:
-    segment_ends = jnp.cumsum(group_sizes, dtype=jnp.int32)
-    segment_starts = jnp.concatenate((jnp.array([0], dtype=segment_ends.dtype), segment_ends[:-1]))
-    positions = jnp.arange(total_size, dtype=jnp.int32)
-    expert_index = jnp.searchsorted(segment_ends, positions, side="right")
-    # Explicitly clip overflow positions to the last segment rather than
-    # depending on implicit out-of-bounds `jnp.take` behavior. Those clipped
-    # positions will have local_rank >= accepted, so they are masked out.
-    expert_index = jnp.minimum(expert_index, group_sizes.shape[0] - 1)
-    local_rank = positions - segment_starts[expert_index]
-    accepted = accepted_group_sizes[expert_index]
-    return local_rank < accepted
-
-
-def _compact_by_keep_mask(inputs: jax.Array, keep_mask: Bool[Array, "T"]) -> jax.Array:
-    total_size = inputs.shape[0]
-    positions = jnp.arange(total_size, dtype=jnp.int32)
-    sort_key = jnp.where(keep_mask, positions, positions + total_size)
-    compacted = _sort_activations(inputs, jnp.argsort(sort_key))
-    valid = positions < jnp.sum(keep_mask.astype(jnp.int32), dtype=jnp.int32)
-    return jnp.where(valid[:, None], compacted, 0)
-
-
-def _expand_from_keep_mask(compacted: jax.Array, keep_mask: Bool[Array, "T"]) -> jax.Array:
-    keep_i32 = keep_mask.astype(jnp.int32)
-    compact_index = jnp.cumsum(keep_i32, dtype=jnp.int32) - 1
-    gathered = jnp.take(compacted, jnp.maximum(compact_index, 0), axis=0)
-    return jnp.where(keep_mask[:, None], gathered, 0)
-
-
-def _moe_mlp_ep_ring_local(
-    x_local: Float[Array, "TL D"],
-    selected_experts_local: Int[Array, "TL K"],
-    combine_weights_local: Float[Array, "TL K"],
-    moe_w13_local: Float[Array, "EL D I2"],
-    moe_w2_local: Float[Array, "EL I D"],
-    *,
-    activation_fn: Callable[[jax.Array], jax.Array],
-    num_experts: int,
-    capacity_factor: float,
-) -> tuple[Float[Array, "TL D"], Int[Array, ""]]:
-    """Ring-style EP routed path: all-gather dispatch + psum-scatter collect."""
-    # #2710 ring EP strategy: gather tokens and their selected-expert routing
-    # assignments across expert shards, then psum-scatter back to local tokens.
-    with jax.named_scope("gather"):
-        x_global = jax.lax.all_gather(x_local, "expert", tiled=True)
-        selected_experts_global = jax.lax.all_gather(selected_experts_local, "expert", tiled=True)
-        combine_weights_global = jax.lax.all_gather(combine_weights_local, "expert", tiled=True)
-
-        tokens = x_global.shape[0]
-        topk = selected_experts_global.shape[1]
-        assignments = tokens * topk
-        expert_flat = selected_experts_global.reshape(assignments)
-        weight_flat = combine_weights_global.reshape(assignments)
-
-        local_experts = moe_w13_local.shape[0]
-        if num_experts % local_experts != 0:
-            raise ValueError(
-                f"num_experts={num_experts} must be divisible by local expert count={local_experts} in EP mode"
-            )
-
-        ep_size = num_experts // local_experts
-        local_capacity = int(math.ceil(capacity_factor * assignments / ep_size))
-        local_capacity = max(local_experts, local_capacity)
-
-        expert_axis = jax.lax.axis_index("expert")
-        expert_start = expert_axis * local_experts
-        local_expert = expert_flat - expert_start
-        local_mask = jnp.logical_and(local_expert >= 0, local_expert < local_experts)
-
-        # Keep only the assignments this shard will execute, ordered by
-        # (local expert id, original flat position). This avoids the global
-        # argsort + fused takes over all assignments that dominated high-EP
-        # shapes, while preserving the grouped layout expected by ragged_dot.
-        local_expert = jnp.where(local_mask, local_expert, 0)
-        # TPU lowers this small-expert count reduction better as a dense
-        # compare+sum than as `bincount`.
-        expert_ids = jnp.arange(local_experts, dtype=jnp.int32)
-        local_mask_i32 = local_mask.astype(jnp.int32)
-        counts = jnp.sum(
-            (local_expert[:, None] == expert_ids[None, :]).astype(jnp.int32) * local_mask_i32[:, None],
-            axis=0,
-            dtype=jnp.int32,
-        )
-        accepted_counts = _prefix_cap_counts(counts, capacity=local_capacity)
-        accepted_total = jnp.sum(accepted_counts, dtype=jnp.int32)
-        dropped_local = jnp.sum(counts, dtype=jnp.int32) - accepted_total
-        valid = jnp.arange(local_capacity, dtype=jnp.int32) < accepted_total
-
-        flat_pos = jnp.arange(assignments, dtype=jnp.int32)
-        order_key = local_expert * assignments + flat_pos
-        max_order_key = local_experts * assignments
-        selection_key = jnp.where(local_mask, max_order_key - order_key, -1)
-        _, local_idx = jax.lax.top_k(selection_key, local_capacity)
-
-        token_local = jnp.floor_divide(local_idx, topk)
-        weight_local = jnp.take(weight_flat, local_idx, axis=0).astype(x_local.dtype)
-
-        x_take = jnp.take(x_global, token_local, axis=0)
-        x_dispatch = jnp.where(valid[:, None], x_take, jnp.zeros_like(x_take))
-        x_dispatch = tree_checkpoint_name(x_dispatch, "grug_moe_dispatch_input")
-        weight_dispatch = jnp.where(valid, weight_local, jnp.zeros_like(weight_local))
-    group_sizes = accepted_counts
-    # `local_idx` pads by appending invalid rows at the end; keep GMM segment
-    # boundaries aligned by attributing padding to the final expert segment.
-    group_sizes = group_sizes.at[-1].add(local_capacity - jnp.sum(group_sizes, dtype=jnp.int32))
-
-    with jax.named_scope("moe_up_down"):
-        w13_out = tree_checkpoint_name(ragged_dot(x_dispatch, moe_w13_local, group_sizes), "grug_moe_expert_hidden")
-        moe_dim = moe_w2_local.shape[1]
-        gate, up = jnp.split(w13_out, [moe_dim], axis=-1)
-        out_dispatch = tree_checkpoint_name(
-            ragged_dot(activation_fn(gate) * up, moe_w2_local, group_sizes),
-            "grug_moe_dispatch_output",
-        )
-
-    with jax.named_scope("scatter"):
-        out_global = jnp.zeros_like(x_global).at[token_local].add(out_dispatch * weight_dispatch[:, None], mode="drop")
-        # #2710 ring EP strategy: collect only this shard's token slice after
-        # reducing contributions from experts across the EP mesh.
-        out_local = jax.lax.psum_scatter(out_global, "expert", scatter_dimension=0, tiled=True)
-        dropped_total = jax.lax.psum(dropped_local, ("data", "expert"))
-    return out_local, dropped_total
-
-
-def _moe_mlp_ep_ragged_a2a_local(
-    x_local: jax.Array,
-    selected_experts_local: jax.Array,
-    combine_weights_local: jax.Array,
-    moe_w13_local: jax.Array,
-    moe_w2_local: jax.Array,
-    *,
-    activation_fn: Callable[[jax.Array], jax.Array],
-    num_experts: int,
-    capacity_factor: float,
-) -> tuple[jax.Array, jax.Array]:
-    local_experts = moe_w13_local.shape[0]
-    if num_experts % local_experts != 0:
-        raise ValueError(
-            f"num_experts={num_experts} must be divisible by local expert count={local_experts} in EP mode"
-        )
-
-    shard_id = jax.lax.axis_index("expert")
-    ep_size = num_experts // local_experts
-    tokens_per_shard = x_local.shape[0]
-    topk = selected_experts_local.shape[1]
-    assignments_per_shard = tokens_per_shard * topk
-    local_capacity = int(math.ceil(capacity_factor * assignments_per_shard))
-    local_capacity = max(local_experts, local_capacity)
-    recv_capacity = local_capacity
-
-    with jax.named_scope("dispatch"):
-        sorted_x, sorted_indices, group_sizes = _permute_by_global_expert(
-            x_local,
-            selected_experts_local,
-            num_experts=num_experts,
-        )
-        all_group_sizes = jax.lax.all_gather(group_sizes.astype(jnp.int32), "expert")
-        clipped_group_sizes = _clip_receiver_group_sizes(
-            all_group_sizes,
-            local_expert_size=local_experts,
-            receiver_capacity=local_capacity,
-        )
-        sender_group_sizes = clipped_group_sizes[shard_id]
-        keep_mask = _expert_prefix_keep_mask(
-            group_sizes.astype(jnp.int32),
-            sender_group_sizes,
-            total_size=assignments_per_shard,
-        )
-        sorted_x = _compact_by_keep_mask(sorted_x, keep_mask)
-
-        all_shard_counts = jnp.sum(clipped_group_sizes.reshape(ep_size, ep_size, local_experts), axis=2)
-        input_offsets, send_sizes, output_offsets, recv_sizes = _shard_a2a_params(all_shard_counts, shard_id)
-        dispatch_out_shape = jnp.zeros((recv_capacity, x_local.shape[1]), dtype=x_local.dtype)
-        x_dispatched = jax.lax.ragged_all_to_all(
-            sorted_x,
-            dispatch_out_shape,
-            input_offsets,
-            send_sizes,
-            output_offsets,
-            recv_sizes,
-            axis_name="expert",
-        )
-        x_dispatch, local_sorted_indices, local_group_sizes = _local_permute_from_counts(
-            x_dispatched,
-            clipped_group_sizes,
-            local_expert_size=local_experts,
-            shard_index=shard_id,
-        )
-
-    with jax.named_scope("moe_up_down"):
-        w13_out = ragged_dot(x_dispatch, moe_w13_local, local_group_sizes)
-        moe_dim = moe_w2_local.shape[1]
-        gate, up = jnp.split(w13_out, [moe_dim], axis=-1)
-        out_dispatch = ragged_dot(activation_fn(gate) * up, moe_w2_local, local_group_sizes)
-
-    with jax.named_scope("combine"):
-        local_output = _sort_activations(out_dispatch, jnp.argsort(local_sorted_indices))
-        return_out_shape = jnp.zeros((assignments_per_shard, x_local.shape[1]), dtype=local_output.dtype)
-        return_input_offsets, return_send_sizes, return_output_offsets, return_recv_sizes = _shard_a2a_params(
-            all_shard_counts.T, shard_id
-        )
-        returned = jax.lax.ragged_all_to_all(
-            local_output,
-            return_out_shape,
-            return_input_offsets,
-            return_send_sizes,
-            return_output_offsets,
-            return_recv_sizes,
-            axis_name="expert",
-        )
-        returned = _expand_from_keep_mask(returned, keep_mask)
-        out_local = _unpermute_from_global_expert(
-            returned,
-            sorted_indices,
-            combine_weights_local,
-            tokens_per_shard=tokens_per_shard,
-            topk=topk,
-        ).astype(x_local.dtype)
-        dropped_local = jnp.sum(group_sizes, dtype=jnp.int32) - jnp.sum(sender_group_sizes, dtype=jnp.int32)
-        dropped_total = jax.lax.psum(dropped_local, ("data", "expert"))
-    return out_local, dropped_total
-
-
 @named_call
 def moe_mlp(
     x: Float[Array, "T D"],
@@ -910,7 +155,7 @@ def moe_mlp(
         mesh = _current_mesh()
 
     if isinstance(activation, ActivationFunctionEnum):
-        activation_fn = activation.to_jax_fn()
+        activation_fn: Callable[[jax.Array], jax.Array] = activation.to_jax_fn()
     else:
         activation_fn = activation
