+++ lib/levanter/src/levanter/grug/moe_ep_ragged_all_to_all.py
@@ -1,0 +1,122 @@
+# Copyright The Levanter Authors
+# SPDX-License-Identifier: Apache-2.0
+
+"""Ragged all-to-all expert-parallel Grug MoE backend."""
+
+import math
+from collections.abc import Callable
+
+import jax
+import jax.numpy as jnp
+
+from haliax.nn.ragged_dot import ragged_dot
+from levanter.grug.moe_ep_common import (
+    _clip_receiver_group_sizes,
+    _compact_by_keep_mask,
+    _expand_from_keep_mask,
+    _expert_prefix_keep_mask,
+    _local_permute_from_counts,
+    _permute_by_global_expert,
+    _shard_a2a_params,
+    _sort_activations,
+    _unpermute_from_global_expert,
+)
+
+
+def _moe_mlp_ep_ragged_a2a_local(
+    x_local: jax.Array,
+    selected_experts_local: jax.Array,
+    combine_weights_local: jax.Array,
+    moe_w13_local: jax.Array,
+    moe_w2_local: jax.Array,
+    *,
+    activation_fn: Callable[[jax.Array], jax.Array],
+    num_experts: int,
+    capacity_factor: float,
+) -> tuple[jax.Array, jax.Array]:
+    local_experts = moe_w13_local.shape[0]
+    if num_experts % local_experts != 0:
+        raise ValueError(
+            f"num_experts={num_experts} must be divisible by local expert count={local_experts} in EP mode"
+        )
+
+    shard_id = jax.lax.axis_index("expert")
+    ep_size = num_experts // local_experts
+    tokens_per_shard = x_local.shape[0]
+    topk = selected_experts_local.shape[1]
+    assignments_per_shard = tokens_per_shard * topk
+    local_capacity = int(math.ceil(capacity_factor * assignments_per_shard))
+    local_capacity = max(local_experts, local_capacity)
+    recv_capacity = local_capacity
+
+    with jax.named_scope("dispatch"):
+        sorted_x, sorted_indices, group_sizes = _permute_by_global_expert(
+            x_local,
+            selected_experts_local,
+            num_experts=num_experts,
+        )
+        all_group_sizes = jax.lax.all_gather(group_sizes.astype(jnp.int32), "expert")
+        clipped_group_sizes = _clip_receiver_group_sizes(
+            all_group_sizes,
+            local_expert_size=local_experts,
+            receiver_capacity=local_capacity,
+        )
+        sender_group_sizes = clipped_group_sizes[shard_id]
+        keep_mask = _expert_prefix_keep_mask(
+            group_sizes.astype(jnp.int32),
+            sender_group_sizes,
+            total_size=assignments_per_shard,
+        )
+        sorted_x = _compact_by_keep_mask(sorted_x, keep_mask)
+
+        all_shard_counts = jnp.sum(clipped_group_sizes.reshape(ep_size, ep_size, local_experts), axis=2)
+        input_offsets, send_sizes, output_offsets, recv_sizes = _shard_a2a_params(all_shard_counts, shard_id)
+        dispatch_out_shape = jnp.zeros((recv_capacity, x_local.shape[1]), dtype=x_local.dtype)
+        x_dispatched = jax.lax.ragged_all_to_all(
+            sorted_x,
+            dispatch_out_shape,
+            input_offsets,
+            send_sizes,
+            output_offsets,
+            recv_sizes,
+            axis_name="expert",
+        )
+        x_dispatch, local_sorted_indices, local_group_sizes = _local_permute_from_counts(
+            x_dispatched,
+            clipped_group_sizes,
+            local_expert_size=local_experts,
+            shard_index=shard_id,
+        )
+
+    with jax.named_scope("moe_up_down"):
+        w13_out = ragged_dot(x_dispatch, moe_w13_local, local_group_sizes)
+        moe_dim = moe_w2_local.shape[1]
+        gate, up = jnp.split(w13_out, [moe_dim], axis=-1)
+        out_dispatch = ragged_dot(activation_fn(gate) * up, moe_w2_local, local_group_sizes)
+
+    with jax.named_scope("combine"):
+        local_output = _sort_activations(out_dispatch, jnp.argsort(local_sorted_indices))
+        return_out_shape = jnp.zeros((assignments_per_shard, x_local.shape[1]), dtype=local_output.dtype)
+        return_input_offsets, return_send_sizes, return_output_offsets, return_recv_sizes = _shard_a2a_params(
+            all_shard_counts.T, shard_id
+        )
+        returned = jax.lax.ragged_all_to_all(
+            local_output,
+            return_out_shape,
+            return_input_offsets,
+            return_send_sizes,
+            return_output_offsets,
+            return_recv_sizes,
+            axis_name="expert",
+        )
+        returned = _expand_from_keep_mask(returned, keep_mask)
+        out_local = _unpermute_from_global_expert(
+            returned,
+            sorted_indices,
+            combine_weights_local,
+            tokens_per_shard=tokens_per_shard,
+            topk=topk,
+        ).astype(x_local.dtype)
+        dropped_local = jnp.sum(group_sizes, dtype=jnp.int32) - jnp.sum(sender_group_sizes, dtype=jnp.int32)
+        dropped_total = jax.lax.psum(dropped_local, ("data", "expert"))
+    return out_local, dropped_total
