+++ lib/levanter/src/levanter/grug/moe_ep_ring.py
@@ -1,0 +1,110 @@
+# Copyright The Levanter Authors
+# SPDX-License-Identifier: Apache-2.0
+
+"""Ring expert-parallel Grug MoE backend."""
+
+import math
+from collections.abc import Callable
+
+import jax
+import jax.numpy as jnp
+from haliax.jax_utils import tree_checkpoint_name
+from jaxtyping import Array, Float, Int
+
+from haliax.nn.ragged_dot import ragged_dot
+from levanter.grug.moe_ep_common import _prefix_cap_counts
+
+
+def _moe_mlp_ep_ring_local(
+    x_local: Float[Array, "TL D"],
+    selected_experts_local: Int[Array, "TL K"],
+    combine_weights_local: Float[Array, "TL K"],
+    moe_w13_local: Float[Array, "EL D I2"],
+    moe_w2_local: Float[Array, "EL I D"],
+    *,
+    activation_fn: Callable[[jax.Array], jax.Array],
+    num_experts: int,
+    capacity_factor: float,
+) -> tuple[Float[Array, "TL D"], Int[Array, ""]]:
+    """Ring-style EP routed path: all-gather dispatch + psum-scatter collect."""
+    # #2710 ring EP strategy: gather tokens and their selected-expert routing
+    # assignments across expert shards, then psum-scatter back to local tokens.
+    with jax.named_scope("gather"):
+        x_global = jax.lax.all_gather(x_local, "expert", tiled=True)
+        selected_experts_global = jax.lax.all_gather(selected_experts_local, "expert", tiled=True)
+        combine_weights_global = jax.lax.all_gather(combine_weights_local, "expert", tiled=True)
+
+        tokens = x_global.shape[0]
+        topk = selected_experts_global.shape[1]
+        assignments = tokens * topk
+        expert_flat = selected_experts_global.reshape(assignments)
+        weight_flat = combine_weights_global.reshape(assignments)
+
+        local_experts = moe_w13_local.shape[0]
+        if num_experts % local_experts != 0:
+            raise ValueError(
+                f"num_experts={num_experts} must be divisible by local expert count={local_experts} in EP mode"
+            )
+
+        ep_size = num_experts // local_experts
+        local_capacity = int(math.ceil(capacity_factor * assignments / ep_size))
+        local_capacity = max(local_experts, local_capacity)
+
+        expert_axis = jax.lax.axis_index("expert")
+        expert_start = expert_axis * local_experts
+        local_expert = expert_flat - expert_start
+        local_mask = jnp.logical_and(local_expert >= 0, local_expert < local_experts)
+
+        # Keep only the assignments this shard will execute, ordered by
+        # (local expert id, original flat position). This avoids the global
+        # argsort + fused takes over all assignments that dominated high-EP
+        # shapes, while preserving the grouped layout expected by ragged_dot.
+        local_expert = jnp.where(local_mask, local_expert, 0)
+        # TPU lowers this small-expert count reduction better as a dense
+        # compare+sum than as `bincount`.
+        expert_ids = jnp.arange(local_experts, dtype=jnp.int32)
+        local_mask_i32 = local_mask.astype(jnp.int32)
+        counts = jnp.sum(
+            (local_expert[:, None] == expert_ids[None, :]).astype(jnp.int32) * local_mask_i32[:, None],
+            axis=0,
+            dtype=jnp.int32,
+        )
+        accepted_counts = _prefix_cap_counts(counts, capacity=local_capacity)
+        accepted_total = jnp.sum(accepted_counts, dtype=jnp.int32)
+        dropped_local = jnp.sum(counts, dtype=jnp.int32) - accepted_total
+        valid = jnp.arange(local_capacity, dtype=jnp.int32) < accepted_total
+
+        flat_pos = jnp.arange(assignments, dtype=jnp.int32)
+        order_key = local_expert * assignments + flat_pos
+        max_order_key = local_experts * assignments
+        selection_key = jnp.where(local_mask, max_order_key - order_key, -1)
+        _, local_idx = jax.lax.top_k(selection_key, local_capacity)
+
+        token_local = jnp.floor_divide(local_idx, topk)
+        weight_local = jnp.take(weight_flat, local_idx, axis=0).astype(x_local.dtype)
+
+        x_take = jnp.take(x_global, token_local, axis=0)
+        x_dispatch = jnp.where(valid[:, None], x_take, jnp.zeros_like(x_take))
+        x_dispatch = tree_checkpoint_name(x_dispatch, "grug_moe_dispatch_input")
+        weight_dispatch = jnp.where(valid, weight_local, jnp.zeros_like(weight_local))
+    group_sizes = accepted_counts
+    # `local_idx` pads by appending invalid rows at the end; keep GMM segment
+    # boundaries aligned by attributing padding to the final expert segment.
+    group_sizes = group_sizes.at[-1].add(local_capacity - jnp.sum(group_sizes, dtype=jnp.int32))
+
+    with jax.named_scope("moe_up_down"):
+        w13_out = tree_checkpoint_name(ragged_dot(x_dispatch, moe_w13_local, group_sizes), "grug_moe_expert_hidden")
+        moe_dim = moe_w2_local.shape[1]
+        gate, up = jnp.split(w13_out, [moe_dim], axis=-1)
+        out_dispatch = tree_checkpoint_name(
+            ragged_dot(activation_fn(gate) * up, moe_w2_local, group_sizes),
+            "grug_moe_dispatch_output",
+        )
+
+    with jax.named_scope("scatter"):
+        out_global = jnp.zeros_like(x_global).at[token_local].add(out_dispatch * weight_dispatch[:, None], mode="drop")
+        # #2710 ring EP strategy: collect only this shard's token slice after
+        # reducing contributions from experts across the EP mesh.
+        out_local = jax.lax.psum_scatter(out_global, "expert", scatter_dimension=0, tiled=True)
+        dropped_total = jax.lax.psum(dropped_local, ("data", "expert"))
+    return out_local, dropped_total
